requirements.txt
# Core dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
h5py>=3.6.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
tqdm>=4.62.0

# Additional
numpy-quaternion>=2022.4.1
PyYAML>=6.0
